lib/seqcmp: Use range-based for loop
As it stands now, we are performing pointer arithmetic on each loop
iteration, adding to the starting address of the passed array its size
in the condition expression. In a range-based for loop, that bound is
computed just once, and range-based for loops are also nicer, so
switching to one looks win-win.